Backend/requirements.txt
# Ultralytics requirements
# Usage: pip install -r requirements.txt
# python -m pip install -r requirements.txt --upgrade
# Base ----------------------------------------
Flask 
flask-cors 
ultralytics 
opencv-python 
numpy 
Pillow